Remove our copies of FileThingpediaClient
The cheatsheet code, which was not moved to the thingpedia
library, is now in util/dataset, near the rest of the cheatsheet
code.
